requirements.txt
numpy
pandas
matplotlib

streamlit

open-iris
